v0.7.2: Fix numpy random API and torch dtype conversion issues
Critical fixes:
- Remove invalid dtype argument from np.random.rand/randn/exponential
  (these NumPy functions don't accept dtype parameter)
- Fix torch backend dtype conversion for zeros/ones/arange/asarray
  (convert np.float64 to torch.float64 properly)
- Fix haar_fluctuation lags dtype (should be int, not float)

Testing improvements:
- Backend consistency tests now check BOTH numpy and torch backends
- All 79 tests pass with both backends

This fixes TypeError on pyodide and other environments where the
invalid dtype argument was rejected.

+++ scaleinvariance/analysis/haar_fluctuation.py
@@ -97,7 +97,7 @@ def haar_fluctuation_analysis(data, order=1, max_sep=None, axis=0, lags='powers 
         mean_absolute_haar_fluctuation = np.nanmean(abs_conv)
         haar_flucs.append(mean_absolute_haar_fluctuation)
 
-    return np.array(lags, dtype=np.float64), np.array(haar_flucs, dtype=np.float64)
+    return np.array(lags, dtype=np.int64), np.array(haar_flucs, dtype=np.float64)
 
 
 def haar_fluctuation_hurst(data, min_sep=None, max_sep=None, axis=0, return_fit=False):
